backend/scaning-service/scanner/router/nmap_scan.py
import nmap

scanner = nmap.PortScanner()
target = "scanme.nmap.org"
scanner.scan(target)
for host in scanner.all_hosts():
    print("Host: ", host)
    print("State: ", scanner[host].state())
    for proto in scanner[host].all_protocols():
        print("Protocol: ", proto)
        ports = scanner[host][proto].keys()
        for port in ports:
            print("Port: ", port, "State: ", scanner[host][proto][port]['state'])
