backend/scaning-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pyjwt
bcrypt
xmltodict
requests

